Move the audio drivers out of the audio target
The drivers are the last thing making audio depend on the application.
They drive the sequencer and the main window directly -- seq->process(),
seq->score(), mscore->midiNoteReceived() and so on, some 17 methods
between them -- so they are the application's glue to the audio
backends rather than part of a synthesis library. mscoreapp already
reached into them for driver.h, pa.h and pm.h.

Compile them into mscoreapp, and keep them together in audiodrivers/
rather than under audio/, so that the split is visible.

Inverting this instead, by extracting interfaces for Seq and MuseScore,
looks possible on paper: iplaypanel.h already does it for the play
panel. It is not attempted here. It would touch the real time audio
callback in six platform specific files that no macOS build compiles,
so CI would be the first thing to ever try it.

With this, audio needs nothing from mscoreapp: only preferences,
extension, icons and globals, which all have targets of their own now.
