+++ mscore/musescore.cpp
@@ -69,7 +69,7 @@
 #include "workspacecombobox.h"
 #include "zoombox.h"
 
-#include "audio/drivers/driver.h"
+#include "audiodrivers/driver.h"
 
 #ifdef USE_LAME
 #include "audio/exports/exportmp3.h"
